+++ config.py
@@ -123,4 +123,4 @@ DEFAULT_UPDATE_SQL = _env_bool("DEFAULT_UPDATE_SQL", False)
 # PROJECT VERSION
 # =========================
 
-PROJECT_VERSION = "0.6.8"
+PROJECT_VERSION = "0.6.9"